MSI #234 Split indexDataProvider class into two separate classes
